Further fine tuning of test structure and test fixes
Make the directory structure more intuitive and fix tests that are failing due to the changed structure
