+++ .github/workflows/README.md
@@ -75,26 +75,28 @@ Reusable workflow called by `helm-push.yaml` for each matrix entry. Accepts:
 
 Runs directory-based **no-secrets** pytest for each changed agent project. Check names in the PR UI: **`tests / corto`**, **`tests / lungo`**, **`tests / recruiter`**.
 
-LLM tests live under `tests/integration_llm/` and are **local-only** (not run in CI) until project leadership re-enables them.
+LLM tests live under `tests/integration/llm/` and are **local-only** (not run in CI) until project leadership re-enables them.
 
 ### Test directories
 
 | Directory | Meaning | CI |
 |-----------|---------|-----|
 | `tests/unit/` | Mocks only | Yes |
-| `tests/live/` | Subprocess uvicorn/A2A (lungo only) | Yes |
-| `tests/integration/` | Docker-compose session; no LLM | Yes |
-| `tests/integration_llm/` | Docker + LLM credentials | No (local manual) |
+| `tests/integration/general/` | Docker-compose session; no live webserver; no LLM | Yes |
+| `tests/integration/live/` | Subprocess uvicorn/A2A HTTP; no LLM | Yes |
+| `tests/integration/llm/` | LLM credentials (may also need docker and/or live server) | No (local manual) |
+
+Shared session fixtures and docker helpers live at `tests/integration/conftest.py` and `tests/integration/helpers/`.
 
 ### Per-project CI pytest paths
 
 | Project | `test_paths` |
 |---------|--------------|
-| corto | `tests/unit tests/integration` |
-| lungo | `tests/unit tests/live tests/integration` |
-| recruiter | `tests/unit tests/integration` |
+| corto | `tests/unit tests/integration/general` |
+| lungo | `tests/unit tests/integration/general tests/integration/live` |
+| recruiter | `tests/unit tests/integration/general tests/integration/live` |
 
-CI always passes explicit paths (never bare `pytest`, which would collect `integration_llm/` via `testpaths = ["tests"]`).
+CI always passes explicit paths (never bare `pytest tests/integration`, which would collect `integration/llm/` via `testpaths = ["tests"]`).
 
 The reusable job sets `WORKFLOW_API_KEY` for lungo live tests (not a repository secret).
 
@@ -102,15 +104,15 @@ Local commands:
 
 ```bash
 # CI-equivalent (no secrets)
-cd coffeeAGNTCY/coffee_agents/corto && uv run pytest tests/unit tests/integration -q
-cd coffeeAGNTCY/coffee_agents/lungo && uv run pytest tests/unit tests/live tests/integration -q
-cd coffeeAGNTCY/coffee_agents/recruiter && uv run pytest tests/unit tests/integration -q
+cd coffeeAGNTCY/coffee_agents/corto && uv run pytest tests/unit tests/integration/general -q
+cd coffeeAGNTCY/coffee_agents/lungo && uv run pytest tests/unit tests/integration/general tests/integration/live -q
+cd coffeeAGNTCY/coffee_agents/recruiter && uv run pytest tests/unit tests/integration/general tests/integration/live -q
 
 # LLM (local only, needs .env)
-cd coffeeAGNTCY/coffee_agents/lungo && uv run pytest tests/integration_llm -q
+cd coffeeAGNTCY/coffee_agents/lungo && uv run pytest tests/integration/llm -q
 ```
 
-Do not run `pytest tests/` as a single full-suite invocation when both `integration/` and `integration_llm/` exist — session fixtures may load twice via `pytest_plugins`.
+Do not run bare `pytest tests/integration` without explicit subpaths — that would collect `integration/llm/`.
 
 ### Branch protection
 
@@ -125,7 +127,7 @@ Do not run `pytest tests/` as a single full-suite invocation when both `integrat
 
 ### Re-enabling LLM CI (future)
 
-When policy allows, add a separate workflow job that runs `pytest tests/integration_llm` with secrets on trusted triggers. The directory layout is already in place.
+When policy allows, add a separate workflow job that runs `pytest tests/integration/llm` with secrets on trusted triggers. The directory layout is already in place.
 
 ### workflow_call inputs
 
@@ -141,7 +143,7 @@ Runs `pytest` via `uv` for a single project directory and explicit path list.
 | Input | Description |
 |-------|-------------|
 | `project_dir` | Path to the project directory to test (required) |
-| `test_paths` | Space-separated pytest paths, e.g. `tests/unit tests/integration` (required) |
+| `test_paths` | Space-separated pytest paths, e.g. `tests/unit tests/integration/general` (required) |
 | `pip_overrides` | PEP 508 specs (one per line) forced into the lock |
 | `pip_constraints` | Constraint lines applied during resolution |
 | `docker_overrides` | Lines `service=image[:tag]` to patch docker-compose service images |
