+++ .github/workflows/test-reusable.yaml
@@ -8,7 +8,7 @@ on:
         required: true
         type: string
       test_paths:
-        description: 'Space-separated pytest paths (e.g. tests/unit tests/integration).'
+        description: 'Space-separated pytest paths (e.g. tests/unit tests/integration/general tests/integration/live).'
         required: true
         type: string
       pip_overrides:
