+++ coffeeAGNTCY/coffee_agents/lungo/pyproject.toml
@@ -29,6 +29,7 @@ dependencies = [
     "langchain-litellm==0.3.5",
     "llama-index-llms-litellm==0.6.3",
     "llama-index-core==0.14.9",
+    "griffe<2",  # banks/llama-index import griffe 1.x API; 2.x is a meta-package without it
     "litellm==1.81.1",
     "google-adk==1.24.0",
     "slima2a==0.3.0",
