+++ coffeeAGNTCY/coffee_agents/lungo/tests/integration/conftest.py
@@ -23,14 +23,14 @@ from pathlib import Path
 import config.config  # noqa: F401 # Note: imports config.config to set environment variables.
 import pytest
 from fastapi.testclient import TestClient
-from tests.integration.docker_helpers import (
+from tests.integration.helpers.docker_helpers import (
     down,
     ensure_compose_profile,
     remove_container_if_exists,
     up,
     wait_for_tcp_port,
 )
-from tests.integration.process_helper import ProcessRunner
+from tests.integration.helpers.process_helper import ProcessRunner
 
 LUNGO_DIR = Path(__file__).resolve().parents[2]
 
@@ -38,8 +38,6 @@ LUNGO_DIR = Path(__file__).resolve().parents[2]
 ensure_compose_profile("observability")
 
 _otel_shutdown_done = False
-
-_auction_shared_initialized_by_integration = False
 
 AGENTS = {
     # auction agents
@@ -111,6 +109,12 @@ def _purge_modules(prefixes, keep=None):
                  and not any(m.startswith(k + ".") for k in keep)]
     for m in to_delete:
         sys.modules.pop(m, None)
+
+
+def _purge_farm_card_modules():
+    for m in list(sys.modules):
+        if m.startswith("agents.farms.") and m.endswith(".card"):
+            sys.modules.pop(m, None)
 
 
 def _save_modules(prefixes):
@@ -405,70 +409,45 @@ def agents_up(request, transport_config):
 # ---------------- http client ----------------
 
 @pytest.fixture
-def auction_supervisor_client(transport_config, monkeypatch, request):
-    global _auction_shared_initialized_by_integration
-
+def auction_supervisor_client(transport_config, monkeypatch):
     for k, v in _base_env().items():
         monkeypatch.setenv(k, str(v))
     for k, v in transport_config.items():
         monkeypatch.setenv(k, v)
-    prefixes = ["agents.supervisors.auction", "config.config"]
-    pin_shared = request.node.get_closest_marker("no_pin_auction_shared") is None
 
-    # First integration use: drop shared imported during unit-test collection
-    if not _auction_shared_initialized_by_integration:
-        sys.modules.pop("agents.supervisors.auction.graph.shared", None)
-        keep = []
-        import importlib
-
-        import config.config as config_mod
-
-        importlib.reload(config_mod)
-        for m in list(sys.modules):
-            if m.startswith("agents.farms.") and m.endswith(".card"):
-                sys.modules.pop(m, None)
-    else:
-        keep = ["agents.supervisors.auction.graph.shared"] if pin_shared else []
-
-    # Collection of unit tests may import recruiter.shared + A2AClientFactory
-    # before auction integration runs; drop it so SLIM native state is not shared.
-    for m in list(sys.modules):
-        if m.startswith(
-            ("agents.supervisors.recruiter", "agents.supervisors.logistics")
-        ):
-            sys.modules.pop(m, None)
-
-    if not pin_shared:
-        for m in list(sys.modules):
-            if m.startswith("agents.farms.") and m.endswith(".card"):
-                sys.modules.pop(m, None)
-
-    reloaded = None
+    prefixes = [
+        "agents.supervisors.auction",
+        "common.a2a_transport_config",
+        "config.config",
+    ]
+    saved = _save_modules(prefixes)
     try:
-        _purge_modules(prefixes, keep=keep)
+        # Unit tests may import recruiter/logistics shared + A2AClientFactory before
+        # integration runs; drop them so SLIM native state is not shared.
+        for m in list(sys.modules):
+            if m.startswith(
+                ("agents.supervisors.recruiter", "agents.supervisors.logistics")
+            ):
+                sys.modules.pop(m, None)
+
+        # Farm cards bake SLIM_SERVER/NATS_SERVER into AgentInterface URLs at import
+        # time; drop them so each test reloads with loopback from _base_env().
+        _purge_farm_card_modules()
+        _purge_modules(prefixes)
+
         import importlib
 
         import agents.supervisors.auction.main as auction_main
 
         importlib.reload(auction_main)
         app = auction_main.app
-
-        # Snapshot the good state only after a successful reload
-        if pin_shared:
-            reloaded = _save_modules(prefixes)
         with TestClient(app) as client:
             _wait_ready(client, "/ready")
             yield client
     finally:
-        # Clear auction modules so the next test can reload cleanly
-        _purge_modules(prefixes, keep=[])
-        if pin_shared and reloaded is not None:
-            _restore_modules(reloaded)
-            _auction_shared_initialized_by_integration = True
-        elif not pin_shared:
-            for m in list(sys.modules):
-                if m.startswith("agents.farms.") and m.endswith(".card"):
-                    sys.modules.pop(m, None)
+        _purge_modules(prefixes)
+        _purge_farm_card_modules()
+        _restore_modules(saved)
 
 
 @pytest.fixture
