+++ coffeeAGNTCY/coffee_agents/lungo/tests/integration/general/_auction_helpers.py
@@ -76,8 +76,8 @@ def get_semantic_similarity(text1, text2):
 
 
 def load_auction_prompt_cases():
-    """Load auction prompt cases from JSON in ``tests/integration/``."""
-    data_file = Path(__file__).resolve().parent.parent / "integration" / "auction_prompt_cases.json"
+    """Load auction prompt cases from JSON in ``tests/integration/general/``."""
+    data_file = Path(__file__).resolve().parent / "auction_prompt_cases.json"
     if not data_file.exists():
         raise FileNotFoundError(f"Prompt cases file not found: {data_file}")
     with data_file.open() as f:
