+++ coffeeAGNTCY/coffee_agents/recruiter/tests/integration/conftest.py
@@ -18,7 +18,7 @@ import httpx
 import pytest
 from dotenv import load_dotenv
 
-from tests.integration.docker_helpers import down, remove_container_if_exists, up
+from tests.integration.helpers.docker_helpers import down, remove_container_if_exists, up
 
 load_dotenv()
